+++ example/TestLinkExampleGenericApi.py
@@ -471,6 +471,13 @@ print "getLastExecutionResult", response
 response = myTestLink.getLastExecutionResult(newTestPlanID_A, 
                                              testcaseid=newTestCaseID_B)
 print "getLastExecutionResult", response
+if not myTLVersion == '<= 1.9.8':
+    # new optional arguments platformid , buildid with TL 1.9.9
+    response = myTestLink.getLastExecutionResult(
+                            newTestPlanID_A, testcaseid=newTestCaseID_AA,
+                            platformid=newPlatFormID_A)
+    print "getLastExecutionResult", response
+    
 response = myTestLink.getExecCountersByBuild(newTestPlanID_A)
 print "getExecCountersByBuild", response
 response = myTestLink.getExecCountersByBuild(newTestPlanID_B)